backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
python-dotenv==1.0.1
supabase==2.10.0
groq==0.11.0
PyMuPDF==1.24.9
python-docx==1.1.2
sentence-transformers==3.3.1
scikit-learn==1.5.2
numpy==2.1.3
pydantic==2.9.2
pydantic-settings==2.6.1
httpx==0.27.2
PyPDF2==3.0.1
